+++ src/assets/js/Router.js
@@ -1,13 +1,13 @@
 export default class Router {
     constructor() {
         this.preRoutes = [
-            '/','/login','/signup','/user', '/reset-password','/adds','/profile','/settings'
+            '/', '/login', '/signup', '/user', '/reset-password', '/adds', '/profile', '/settings'
         ]
         this.routes = []
     }
 
-    get(uri, callback) { 
-        
+    get(uri, callback) {
+
         // ensure that the parameters are not empty
         if (!uri || !callback) throw new Error('uri or callback must be given')
 
@@ -29,15 +29,15 @@ export default class Router {
             callback
         }
         this.routes.push(route)
-        
+
         // If route defined then no need to compare all of the routes; imediately call the init() method
-        if(this.preRoutes.includes(uri)) {
+        if (this.preRoutes.includes(uri)) {
             this.init()
             this.routes = []
         }
     }
 
-    getRoute() {}
+    getRoute() { }
 
     init() {
         this.routes.some(route => {
@@ -49,7 +49,7 @@ export default class Router {
 
                 let req = { path } // i'll also explain this code below
                 return route.callback.call(this, req)
-            }
+            } else dispatchEvent(new CustomEvent("error", { detail: { path: path, err: '404' } }))
         })
     }
 }
